Ver. 1.0  .................................................................

+++ src/make_circuitfile/make_cir_last.cpp
@@ -18,7 +18,7 @@ using namespace std;
 int make_cir_last(vector<ele_unit> &element, vector<string> &data_cir, vector<string> &arg_arr){
     stringstream ciroutname;
     ciroutname << arg_arr[1] << "_out" << ".cir";
-    for(int i = 0; i < arg_arr.size(); i++){
+    for(size_t i = 0; i < arg_arr.size(); i++){
         if(arg_arr[i] == "-o"){
             ciroutname.clear();
             ciroutname.str("");
@@ -40,7 +40,7 @@ int make_cir_last(vector<ele_unit> &element, vector<string> &data_cir, vector<st
         cerr << "file not error3!"  << endl;
         return 0;
     }        
-    for (int x = 0; x < element.size(); x++) {
+    for (size_t x = 0; x < element.size(); x++) {
         liness.clear();
         liness.str("");
         y = element[x].line_num;
@@ -104,7 +104,7 @@ int make_cir_last(vector<ele_unit> &element, vector<string> &data_cir, vector<st
         }
         transform(element[x].element.begin(), element[x].element.end(), element[x].element.begin(), ::toupper); 
     }
-    for (int x = 0; x < data_cir_copy.size(); x++) {     //内容を書き換えたdata_cir_copy(バッファ)を全て書く
+    for (size_t x = 0; x < data_cir_copy.size(); x++) {     //内容を書き換えたdata_cir_copy(バッファ)を全て書く
         if(data_cir_copy[x].find(".file") != string::npos || data_cir_copy[x].find(".FILE") != string::npos || data_cir_copy[x].find(".File") != string::npos){
             fpin << ".FILE CIRCUIT" <<  getpid() << ".CSV" << endl;
         }
