+++ src/read_file/sub_unit.cpp
@@ -16,12 +16,12 @@
 using namespace std;
 
 string sub_unit(string& main){
-    int find_sub1 = 0, find_sub2 = 0;
+    size_t find_sub1 = 0, find_sub2 = 0;
     int i = 0;
     find_sub1 = main.find("f");
     find_sub2 = main.find("F");
     if(find_sub1 != string::npos || find_sub2 != string::npos){
-        i = max({find_sub1,find_sub2});
+        i = max({(int)find_sub1,(int)find_sub2});
         main.insert(i," ");     
         return "f";
     }
@@ -29,7 +29,7 @@ string sub_unit(string& main){
     find_sub1 = main.find("p");
     find_sub2 = main.find("P");
     if(find_sub1 != string::npos || find_sub2 != string::npos){
-        i = max({find_sub1,find_sub2});
+        i = max({(int)find_sub1,(int)find_sub2});
         main.insert(i," ");     
         return "p";
     }
@@ -37,7 +37,7 @@ string sub_unit(string& main){
     find_sub1 = main.find("n");
     find_sub2 = main.find("N");
     if(find_sub1 != string::npos || find_sub2 != string::npos){
-        i = max({find_sub1,find_sub2});
+        i = max({(int)find_sub1,(int)find_sub2});
         main.insert(i," ");     
         return "n";
     }
@@ -45,7 +45,7 @@ string sub_unit(string& main){
     find_sub1 = main.find("u");
     find_sub2 = main.find("U");
     if(find_sub1 != string::npos || find_sub2 != string::npos){
-        i = max({find_sub1,find_sub2});
+        i = max({(int)find_sub1,(int)find_sub2});
         main.insert(i," ");     
         return "u";
     }
@@ -53,7 +53,7 @@ string sub_unit(string& main){
     find_sub1 = main.find("m");
     find_sub2 = main.find("M");
     if(find_sub1 != string::npos || find_sub2 != string::npos){
-        i = max({find_sub1,find_sub2});
+        i = max({(int)find_sub1,(int)find_sub2});
         main.insert(i," ");     
         return "m";
     }
@@ -61,7 +61,7 @@ string sub_unit(string& main){
     find_sub1 = main.find("k");
     find_sub2 = main.find("K");
     if(find_sub1 != string::npos || find_sub2 != string::npos){
-        i = max({find_sub1,find_sub2});
+        i = max({(int)find_sub1,(int)find_sub2});
         main.insert(i," ");     
         return "k";
     }
@@ -69,7 +69,7 @@ string sub_unit(string& main){
     find_sub1 = main.find("meg");
     find_sub2 = main.find("MEG");
     if(find_sub1 != string::npos || find_sub2 != string::npos){
-        i = max({find_sub1,find_sub2});
+        i = max({(int)find_sub1,(int)find_sub2});
         main.insert(i," ");     
         return "meg";
     }
@@ -77,7 +77,7 @@ string sub_unit(string& main){
     find_sub1 = main.find("x");
     find_sub2 = main.find("X");
     if(find_sub1 != string::npos || find_sub2 != string::npos){
-        i = max({find_sub1,find_sub2});
+        i = max({(int)find_sub1,(int)find_sub2});
         main.insert(i," ");     
         return "x";
     }
@@ -85,7 +85,7 @@ string sub_unit(string& main){
     find_sub1 = main.find("g");
     find_sub2 = main.find("G");
     if(find_sub1 != string::npos || find_sub2 != string::npos){
-        i = max({find_sub1,find_sub2});
+        i = max({(int)find_sub1,(int)find_sub2});
         main.insert(i," ");     
         return "g";
     }
@@ -93,7 +93,7 @@ string sub_unit(string& main){
     find_sub1 = main.find("t");
     find_sub2 = main.find("T");
     if(find_sub1 != string::npos || find_sub2 != string::npos){
-        i = max({find_sub1,find_sub2});
+        i = max({(int)find_sub1,(int)find_sub2});
         main.insert(i," ");     
         return "t";
     }
